Revert "Checkout: Centralize logic that determines which contact form to use (#47080)"
This reverts commit 957c7514ad28fe41d84f0e0845be8d9625a5713e.
