+++ client/my-sites/checkout/composite-checkout/components/wp-checkout.js
@@ -23,11 +23,11 @@ import {
 	useTotal,
 } from '@automattic/composite-checkout';
 import debugFactory from 'debug';
-import { useShoppingCart } from '@automattic/shopping-cart';
 
 /**
  * Internal dependencies
  */
+import { areDomainsInLineItems, isLineItemADomain } from '../hooks/has-domains';
 import useCouponFieldState from '../hooks/use-coupon-field-state';
 import useUpdateCartLocationWhenPaymentMethodChanges from '../hooks/use-update-cart-location-when-payment-method-changes';
 import WPCheckoutOrderReview from './wp-checkout-order-review';
@@ -46,15 +46,10 @@ import {
 	getSignupEmailValidationResult,
 	getGSuiteValidationResult,
 } from 'calypso/my-sites/checkout/composite-checkout/contact-validation';
+import { isGSuiteProductSlug } from 'calypso/lib/gsuite';
 import { needsDomainDetails } from 'calypso/my-sites/checkout/composite-checkout/payment-method-helpers';
 import { login } from 'calypso/lib/paths';
 import config from 'calypso/config';
-import getContactDetailsType from '../lib/get-contact-details-type';
-import {
-	hasGoogleApps,
-	hasDomainRegistration,
-	hasTransferProduct,
-} from 'calypso/lib/cart-values/cart-items';
 
 const debug = debugFactory( 'calypso:composite-checkout:wp-checkout' );
 
@@ -62,12 +57,12 @@ const ContactFormTitle = () => {
 	const translate = useTranslate();
 	const isActive = useIsStepActive();
 	const isComplete = useIsStepComplete();
-	const { responseCart } = useShoppingCart();
-	const areThereDomainProductsInCart =
-		hasDomainRegistration( responseCart ) || hasTransferProduct( responseCart );
-	const isGSuiteInCart = hasGoogleApps( responseCart );
+	const [ items ] = useLineItems();
+	const isGSuiteInCart = items.some( ( item ) =>
+		isGSuiteProductSlug( item.wpcom_meta?.product_slug )
+	);
 
-	if ( areThereDomainProductsInCart ) {
+	if ( areDomainsInLineItems( items ) ) {
 		return ! isActive && isComplete
 			? translate( 'Contact information' )
 			: translate( 'Enter your contact information' );
@@ -118,12 +113,13 @@ export default function WPCheckout( {
 	const onEvent = useEvents();
 
 	const [ items ] = useLineItems();
-	const areThereDomainProductsInCart =
-		hasDomainRegistration( responseCart ) || hasTransferProduct( responseCart );
-	const isGSuiteInCart = hasGoogleApps( responseCart );
-
-	const contactDetailsType = getContactDetailsType( responseCart );
-	const shouldShowContactStep = contactDetailsType !== 'none';
+	const areThereDomainProductsInCart = items.some( isLineItemADomain );
+	const isGSuiteInCart = items.some( ( item ) =>
+		isGSuiteProductSlug( item.wpcom_meta?.product_slug )
+	);
+	const shouldShowContactStep =
+		areThereDomainProductsInCart || isGSuiteInCart || total.amount.value > 0;
+	const shouldShowDomainContactFields = shouldShowContactStep && needsDomainDetails( responseCart );
 	const areDomainDetailsNeededForTransaction = needsDomainDetails( responseCart ) || isGSuiteInCart;
 
 	const contactInfo = useSelect( ( sel ) => sel( 'wpcom' ).getContactInfo() ) || {};
@@ -368,7 +364,7 @@ export default function WPCheckout( {
 										shouldShowContactDetailsValidationErrors
 									}
 									contactValidationCallback={ validateContactDetails }
-									contactDetailsType={ contactDetailsType }
+									shouldShowDomainContactFields={ shouldShowDomainContactFields }
 									isLoggedOutCart={ isLoggedOutCart }
 								/>
 							}
