--- client/my-sites/checkout/composite-checkout/types/contact-details.ts
@@ -1,1 +1,0 @@
-export type ContactDetailsType = 'gsuite' | 'tax' | 'domain' | 'none';
